+++ .vscode/settings.json
@@ -3,5 +3,7 @@
   "deno.lint": true,
   "deno.unstable": false,
   "editor.formatOnSave": true,
-  "editor.defaultFormatter": "denoland.vscode-deno"
+  "[typescript]": {
+    "editor.defaultFormatter": "vscode.typescript-language-features"
+  }
 }
