+++ package.json
@@ -9,11 +9,13 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@axiomhq/axiom-node": "^0.5.0",
     "@headlessui/react": "^1.6.4",
     "classnames": "^2.3.1",
     "next": "12.1.6",
     "next-axiom": "^0.7.0",
     "next-seo": "^5.4.0",
+    "node-hn-api": "^2.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
